Moved the events and attendee into the agenda package in the domain layer
