+++ domain/src/main/java/me/androidbox/domain/agenda/model/Attendee.kt
@@ -1,4 +1,4 @@
-package me.androidbox.domain.authentication.model
+package me.androidbox.domain.agenda.model
 
 data class Attendee(
     val id: Int,
