+++ domain/src/main/java/me/androidbox/domain/agenda/model/Event.kt
@@ -1,4 +1,4 @@
-package me.androidbox.domain.authentication.model
+package me.androidbox.domain.agenda.model
 
 data class Event(
     val id: String,
