+++ domain/src/main/java/me/androidbox/domain/authentication/remote/EventRepository.kt
@@ -2,7 +2,7 @@ package me.androidbox.domain.authentication.remote
 
 import kotlinx.coroutines.flow.Flow
 import me.androidbox.domain.authentication.ResponseState
-import me.androidbox.domain.authentication.model.Event
+import me.androidbox.domain.agenda.model.Event
 
 interface EventRepository {
     fun getEventsFromTimeStamp(startTimeStamp: Long, endTimeStamp: Long): Flow<ResponseState<List<Event>>>
