+++ presentation/src/main/java/me/androidbox/presentation/event/screen/EventScreenEvent.kt
@@ -4,7 +4,7 @@ import me.androidbox.component.agenda.AgendaActionType
 import me.androidbox.component.agenda.AlarmReminderItem
 import me.androidbox.component.agenda.VisitorType
 import me.androidbox.component.event.VisitorInfo
-import me.androidbox.domain.authentication.model.Attendee
+import me.androidbox.domain.agenda.model.Attendee
 import java.time.ZonedDateTime
 
 sealed interface EventScreenEvent {
